+++ postgres/load.ts
@@ -1,11 +1,35 @@
 const fs = require('fs');
 const { Client } = require('pg');
 
+/**
+ * Postgres Host
+ */
+let host = process.env.POSTGRES_HOST;
+if (!host) {
+  host = 'host.docker.internal';
+}
+
+/**
+ * Postgres User
+ */
+let user = process.env.POSTGRES_USER;
+if (!user) {
+  user = 'postgres';
+}
+
+/**
+ * Postgres Password
+ */
+let password = process.env.POSTGRES_PASSWORD;
+if (!password) {
+  password = 'postgres';
+}
+
 async function loadFile() {
   /**
    * Connect to Postgres
    */
-  const client = new Client({ user: '', password: '' });
+  const client = new Client({ host, user, password });
   await client.connect();
 
   /**
@@ -17,7 +41,7 @@ async function loadFile() {
   /**
    * Load Image to Postgres
    */
-  const imgData = fs.readFileSync('../src/img/image-panel.png');
+  const imgData = fs.readFileSync('./image-panel.png');
   await client.query('INSERT INTO images VALUES($1, $2)', ['image-panel', imgData]);
 
   /**