+++ src/constants.ts
@@ -1,23 +1,25 @@
 /**
  * Image Types
  */
-export enum ImageTypes {
+export enum SupportedTypes {
   JPEG = 'image/jpeg',
   GIF = 'image/gif',
   PNG = 'image/png',
   PDF = 'application/pdf',
   HEIC = 'image/heic',
+  MP4 = 'video/mp4',
+  WEBM = 'video/webm',
 }
 
 /**
  * Base64 symbols for Image Types
  */
 export enum ImageTypesSymbols {
-  '/' = ImageTypes.JPEG,
-  'R' = ImageTypes.GIF,
-  'i' = ImageTypes.PNG,
-  'J' = ImageTypes.PDF,
-  'A' = ImageTypes.HEIC,
+  '/' = SupportedTypes.JPEG,
+  'R' = SupportedTypes.GIF,
+  'i' = SupportedTypes.PNG,
+  'J' = SupportedTypes.PDF,
+  'A' = SupportedTypes.HEIC,
 }
 
 /**