+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - Rebuild based on 9.0.0-beta3 (#36)
 - Update PDF loader to start in Docker (#38)
 - Support short Video encoded using Base64 (#39)
-- Add Options for Video/Audio Controls and Auto Play (#40)
+- Support MP3, OGG Audio playback in Base64 (#40)
+- Add Options for Video/Audio Controls and Auto Play (#41)
 
 ## 2.6.0 (2022-04-26)